refactor: inline city-timezones lookup into server action, delete searchCities.ts
