--- lib/directory/types.ts
@@ -1,9 +1,0 @@
-/**
- * City entity used by the directory search flow
- */
-export interface City {
-  city: string;
-  city_ascii: string;
-  admin_name: string;
-  country: string;
-}
