+++ ui/signup/CitySearchDropdown.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { searchCitiesAction } from "@/lib/directory/searchCitiesAction";
-import type { City } from "@/lib/directory/types";
+import type { City } from "@/lib/directory/searchCitiesAction";
 import { withFieldBorderState } from "@/ui/common/forms/styles";
 
 type CityWithFullLabel = City & { fullLabel?: string };
